+ Improve parser-lib API
