+++ examples/R/parser.R
@@ -2,5 +2,5 @@ require("rJava")
 pathToGnParserAssembly='${GnParser_PATH}/gnparser/parser/target/scala-2.11/global-names-parser-assembly-0.1.0-SNAPSHOT.jar'
 .jinit(classpath=pathToGnParserAssembly)
 snp=J('org.globalnames.parser.ScientificNameParser','instance')
-result=snp$fromString("Homo sapiens L.")
-print(snp$renderCompactJson(result))
+result=snp$fromString("Homo sapiens L.")$renderCompactJson
+print(result)
