+++ examples/jruby/parser.rb
@@ -2,5 +2,5 @@ require "java"
 java_import "org.globalnames.parser.ScientificNameParser"
 
 snp = ScientificNameParser.instance
-result = snp.fromString("Homo sapiens L.")
-puts snp.renderCompactJson(result)
+result = snp.fromString("Homo sapiens L.").renderCompactJson
+puts result
