+++ examples/jython/parser.py
@@ -1,5 +1,5 @@
 from org.globalnames.parser import ScientificNameParser
 
 snp = ScientificNameParser.instance()
-result = snp.fromString("Homo sapiens L.")
-print snp.renderCompactJson(result)
+result = snp.fromString("Homo sapiens L.").renderCompactJson()
+print result
